// src/domain/contracts/repos/schema.ts
export type TUfEmi =
  | 'AC'
  | 'AL'
  | 'AM'
  | 'AP'
  | 'BA'
  | 'CE'
  | 'DF'
  | 'ES'
  | 'GO'
  | 'MA'
  | 'MG'
  | 'MS'
  | 'MT'
  | 'PA'
  | 'PB'
  | 'PE'
  | 'PI'
  | 'PR'
  | 'RJ'
  | 'RN'
  | 'RO'
  | 'RR'
  | 'RS'
  | 'SC'
  | 'SE'
  | 'SP'
  | 'TO'
  | 'EX';
// ...
type ItemsChoiceType5 = 'balsa' | 'reboque' | 'vagao' | 'veicTransp';
// ...
type TNFeDoc = 'CNPJ' | 'CPF';
// ...
export interface TNFeInfNFeTransp {
  modFrete: '0' | '1' | '2' | '3' | '4' | '9';
  transporta?: TNFeInfNFeTranspTransporta;
  retTransp?: TNFeInfNFeTranspRetTransp;
  veicTransp?: TNFeInfNFeTranspVeicTransp;
  items?: object[];
  itemsElementName?: ItemsChoiceType5[];
  vol?: TNFeInfNFeTranspVol[];
}
interface TNFeInfNFeTranspVol {
  qVol: string;
  esp: string;
  marca?: string;
  nVol?: string;
  pesoL: string;
  pesoB: string;
  lacres?: TNFeInfNFeTranspVolLacres[];
}
interface TNFeInfNFeTranspVolLacres {
  nLacre: string;
}
interface TNFeInfNFeTranspVeicTransp {
  placa: string;
  UF: string;
  RNTC: string;
}
interface TNFeInfNFeTranspRetTransp {
  vServ: string;
  vBCRet: string;
  pICMSRet: string;
  vICMSRet: string;
  cFOP: string;
  cMunFG: string;
}
interface TNFeInfNFeTranspTransporta {
  CNPJ?: string;
  CPF?: string;
  item: string;
  itemElementName: TNFeDoc;
  xNome: string;
  ie: string;
  xEnder: string;
  xMun: string;
  uf: TUfEmi;
  ufSpecified: boolean;
}